follow-up: working draft of cross-product animation
Credits to GitHub Copilot for putting me on the right track again, and thanks to me for determination writing Vekt.js, and a constant pursue of deepening Maths knowledge found in the Mechanics, and beyond. Finally, this is just a working draft, and rather a conceptual animation as achieving 3D in 2D is not an easy bite to tackle.